requirements.txt
fastapi
python-multipart
fal_client
Pillow
ruff
requests
